Split the DisplayWorkV1 tests into multiple classes
