+++ sbt_common/display/src/test/scala/uk/ac/wellcome/display/models/v1/DisplayPublicationDateSerialisationTest.scala
@@ -1,0 +1,75 @@
+package uk.ac.wellcome.display.models.v1
+
+import com.fasterxml.jackson.databind.ObjectMapper
+import com.google.inject.Guice
+import org.scalatest.FunSpec
+import uk.ac.wellcome.display.models.{DisplaySerialisationTestBase, WorksUtil}
+import uk.ac.wellcome.display.modules.DisplayJacksonModule
+import uk.ac.wellcome.models.{IdentifiedWork, Period}
+import uk.ac.wellcome.test.utils.JsonTestUtil
+
+class DisplayPublicationDateSerialisationTest
+  extends FunSpec
+    with DisplaySerialisationTestBase
+    with JsonTestUtil
+    with WorksUtil {
+  val injector = Guice.createInjector(DisplayJacksonModule)
+
+  val objectMapper = injector.getInstance(classOf[ObjectMapper])
+
+    it("omits the publicationDate field if it is empty") {
+      val work = IdentifiedWork(
+        canonicalId = "arfj5cj4",
+        sourceIdentifier = sourceIdentifier,
+        title = Some("Asking aging armadillos for another appraisal"),
+        publicationDate = None,
+        version = 1
+      )
+      val displayWork = DisplayWorkV1(work)
+
+      val actualJson = objectMapper.writeValueAsString(displayWork)
+      val expectedJson = s"""
+                            |{
+                            |  "type": "Work",
+                            |  "id": "${work.canonicalId}",
+                            |  "title": "${work.title.get}",
+                            |  "creators": [ ],
+                            |  "subjects": [ ],
+                            |  "genres": [ ],
+                            |  "publishers": [],
+                            |  "placesOfPublication": [ ]
+                            |}""".stripMargin
+
+      assertJsonStringsAreEqual(actualJson, expectedJson)
+    }
+
+    it("includes the publicationDate field if it is present on the Work") {
+      val work = IdentifiedWork(
+        canonicalId = "avfpwgrr",
+        sourceIdentifier = sourceIdentifier,
+        title = Some("Ahoy!  Armoured angelfish are attacking the armada!"),
+        publicationDate = Some(Period("1923")),
+        version = 1
+      )
+      val displayWork = DisplayWorkV1(work)
+
+      val actualJson = objectMapper.writeValueAsString(displayWork)
+      val expectedJson = s"""
+                            |{
+                            |  "type": "Work",
+                            |  "id": "${work.canonicalId}",
+                            |  "title": "${work.title.get}",
+                            |  "creators": [ ],
+                            |  "subjects": [ ],
+                            |  "genres": [ ],
+                            |  "publishers": [],
+                            |  "publicationDate": ${period(
+        work.publicationDate.get)},
+                            |  "placesOfPublication": [ ]
+                            |}""".stripMargin
+
+      assertJsonStringsAreEqual(actualJson, expectedJson)
+    }
+
+
+}
