+++ sbt_common/display/src/test/scala/uk/ac/wellcome/display/models/v1/DisplayWorkV1SerialisationTest.scala
@@ -305,289 +305,4 @@ class DisplayWorkV1SerialisationTest
 
     assertJsonStringsAreEqual(actualJson, expectedJson)
   }
-
-  describe("creators") {
-    it("serialises creators with a mixture of agents/organisations/persons") {
-      val work = IdentifiedWork(
-        canonicalId = "v9w6cz66",
-        sourceIdentifier = sourceIdentifier,
-        version = 1,
-        title = Some("Vultures vying for victory"),
-        creators = List(
-          Unidentifiable(Agent("Vivian Violet")),
-          Unidentifiable(Organisation("Verily Volumes")),
-          Unidentifiable(
-            Person(
-              label = "Havelock Vetinari",
-              prefix = Some("Lord Patrician"),
-              numeration = Some("I")))
-        )
-      )
-      val displayWork = DisplayWorkV1(work)
-
-      val actualJson = objectMapper.writeValueAsString(displayWork)
-      val expectedJson = s"""
-        |{
-        |  "type": "Work",
-        |  "id": "${work.canonicalId}",
-        |  "title": "${work.title.get}",
-        |  "creators": [
-        |    ${identifiedOrUnidentifiable(work.creators(0), abstractAgent)},
-        |    ${identifiedOrUnidentifiable(work.creators(1), abstractAgent)},
-        |    ${identifiedOrUnidentifiable(work.creators(2), abstractAgent)}
-        |  ],
-        |  "subjects": [ ],
-        |  "genres": [ ],
-        |  "publishers": [],
-        |  "placesOfPublication": [ ]
-        |}""".stripMargin
-
-      assertJsonStringsAreEqual(actualJson, expectedJson)
-    }
-
-    it("serialises identified creators") {
-      val work = IdentifiedWork(
-        canonicalId = "v9w6cz66",
-        sourceIdentifier = sourceIdentifier,
-        version = 1,
-        title = Some("Vultures vying for victory"),
-        creators = List(
-          Identified(
-            Person(
-              label = "Havelock Vetinari",
-              prefix = Some("Lord Patrician"),
-              numeration = Some("I")),
-            canonicalId = "hgfedcba",
-            identifiers = List(
-              SourceIdentifier(
-                IdentifierSchemes.libraryOfCongressNames,
-                ontologyType = "Organisation",
-                value = "hv"))
-          ),
-          Identified(
-            Organisation(label = "Unseen University"),
-            canonicalId = "abcdefgh",
-            identifiers = List(
-              SourceIdentifier(
-                IdentifierSchemes.libraryOfCongressNames,
-                ontologyType = "Organisation",
-                value = "uu"))
-          ),
-          Identified(
-            Agent(label = "The Librarian"),
-            canonicalId = "blahbluh",
-            identifiers = List(
-              SourceIdentifier(
-                IdentifierSchemes.libraryOfCongressNames,
-                ontologyType = "Organisation",
-                value = "uu"))
-          )
-        )
-      )
-      val displayWork = DisplayWorkV1(work)
-
-      val actualJson = objectMapper.writeValueAsString(displayWork)
-      val expectedJson = s"""
-                             |{
-                             |  "type": "Work",
-                             |  "id": "${work.canonicalId}",
-                             |  "title": "${work.title.get}",
-                             |  "creators": [
-                             |    ${identifiedOrUnidentifiable(
-                              work.creators(0),
-                              abstractAgent)},
-                             |    ${identifiedOrUnidentifiable(
-                              work.creators(1),
-                              abstractAgent)},
-                             |    ${identifiedOrUnidentifiable(
-                              work.creators(2),
-                              abstractAgent)}
-                             |  ],
-                             |  "subjects": [ ],
-                             |  "genres": [ ],
-                             |  "publishers": [],
-                             |  "placesOfPublication": [ ]
-                             |}""".stripMargin
-
-      assertJsonStringsAreEqual(actualJson, expectedJson)
-    }
-  }
-
-  describe("locations") {
-    it("serialises a physical location correctly") {
-      val physicalLocation = PhysicalLocation(
-        locationType = "smeg",
-        label = "a stack of slick slimes"
-      )
-
-      val work = IdentifiedWork(
-        canonicalId = "zm9q6c6h",
-        sourceIdentifier = sourceIdentifier,
-        version = 1,
-        title = Some("A zoo of zebras doing zumba"),
-        items = List(
-          IdentifiedItem(
-            canonicalId = "mhberjwy7",
-            sourceIdentifier = sourceIdentifier,
-            locations = List(physicalLocation)
-          )
-        )
-      )
-      val displayWork =
-        DisplayWorkV1(work, includes = WorksIncludes(items = true))
-
-      val actualJson = objectMapper.writeValueAsString(displayWork)
-      val expectedJson = s"""
-                             |{
-                             |  "type": "Work",
-                             |  "id": "${work.canonicalId}",
-                             |  "title": "${work.title.get}",
-                             |  "creators": [ ],
-                             |  "items": [ ${items(work.items)} ],
-                             |  "subjects": [ ],
-                             |  "genres": [ ],
-                             |  "publishers": [],
-                             |  "placesOfPublication": [ ]
-                             |}""".stripMargin
-
-      assertJsonStringsAreEqual(actualJson, expectedJson)
-    }
-  }
-
-  describe("place of publication") {
-    it("serialises the placesOfPublication field") {
-      val work = IdentifiedWork(
-        canonicalId = "avfpwgrr",
-        sourceIdentifier = sourceIdentifier,
-        title = Some("Ahoy!  Armoured angelfish are attacking the armada!"),
-        placesOfPublication = List(Place("Durmstrang")),
-        version = 1
-      )
-      val displayWork = DisplayWorkV1(work)
-
-      val actualJson = objectMapper.writeValueAsString(displayWork)
-      val expectedJson = s"""
-                             |{
-                             |  "type": "Work",
-                             |  "id": "${work.canonicalId}",
-                             |  "title": "${work.title.get}",
-                             |  "creators": [ ],
-                             |  "subjects": [ ],
-                             |  "genres": [ ],
-                             |  "publishers": [],
-                             |  "placesOfPublication": [
-                             |    {
-                             |      "label": "${work.placesOfPublication.head.label}",
-                             |      "type": "Place"
-                             |    }
-                             |  ]
-                             |
-                             |}""".stripMargin
-
-      assertJsonStringsAreEqual(actualJson, expectedJson)
-    }
-  }
-
-  describe("publication date") {
-    it("omits the publicationDate field if it is empty") {
-      val work = IdentifiedWork(
-        canonicalId = "arfj5cj4",
-        sourceIdentifier = sourceIdentifier,
-        title = Some("Asking aging armadillos for another appraisal"),
-        publicationDate = None,
-        version = 1
-      )
-      val displayWork = DisplayWorkV1(work)
-
-      val actualJson = objectMapper.writeValueAsString(displayWork)
-      val expectedJson = s"""
-                            |{
-                            |  "type": "Work",
-                            |  "id": "${work.canonicalId}",
-                            |  "title": "${work.title.get}",
-                            |  "creators": [ ],
-                            |  "subjects": [ ],
-                            |  "genres": [ ],
-                            |  "publishers": [],
-                            |  "placesOfPublication": [ ]
-                            |}""".stripMargin
-
-      assertJsonStringsAreEqual(actualJson, expectedJson)
-    }
-
-    it("includes the publicationDate field if it is present on the Work") {
-      val work = IdentifiedWork(
-        canonicalId = "avfpwgrr",
-        sourceIdentifier = sourceIdentifier,
-        title = Some("Ahoy!  Armoured angelfish are attacking the armada!"),
-        publicationDate = Some(Period("1923")),
-        version = 1
-      )
-      val displayWork = DisplayWorkV1(work)
-
-      val actualJson = objectMapper.writeValueAsString(displayWork)
-      val expectedJson = s"""
-                            |{
-                            |  "type": "Work",
-                            |  "id": "${work.canonicalId}",
-                            |  "title": "${work.title.get}",
-                            |  "creators": [ ],
-                            |  "subjects": [ ],
-                            |  "genres": [ ],
-                            |  "publishers": [],
-                            |  "publicationDate": ${period(
-                              work.publicationDate.get)},
-                            |  "placesOfPublication": [ ]
-                            |}""".stripMargin
-
-      assertJsonStringsAreEqual(actualJson, expectedJson)
-    }
-  }
-
-  describe("publishers") {
-    it(
-      "includes the publishers field with a mixture of agents/organisations/persons") {
-      val work = IdentifiedWork(
-        canonicalId = "v9w6cz66",
-        sourceIdentifier = sourceIdentifier,
-        version = 1,
-        title = Some("Vultures vying for victory"),
-        publishers = List(
-          Unidentifiable(Agent("Vivian Violet")),
-          Unidentifiable(Organisation("Verily Volumes")),
-          Unidentifiable(
-            Person(
-              label = "Havelock Vetinari",
-              prefix = Some("Lord Patrician"),
-              numeration = Some("I")))
-        )
-      )
-      val displayWork = DisplayWorkV1(work)
-
-      val actualJson = objectMapper.writeValueAsString(displayWork)
-      val expectedJson = s"""
-                            |{
-                            |  "type": "Work",
-                            |  "id": "${work.canonicalId}",
-                            |  "title": "${work.title.get}",
-                            |  "creators": [ ],
-                            |  "subjects": [ ],
-                            |  "genres": [ ],
-                            |  "publishers": [
-                            |    ${identifiedOrUnidentifiable(
-                              work.publishers(0),
-                              abstractAgent)},
-                            |    ${identifiedOrUnidentifiable(
-                              work.publishers(1),
-                              abstractAgent)},
-                            |    ${identifiedOrUnidentifiable(
-                              work.publishers(2),
-                              abstractAgent)}
-                            |  ],
-                            |  "placesOfPublication": [ ]
-                            |}""".stripMargin
-
-      assertJsonStringsAreEqual(actualJson, expectedJson)
-    }
-  }
 }
